Move WasapiRealTimeThread into an opaque `unique_ptr`
This gets *all* the Windows include nonsense out of the header, ensuring that:
- The header iteslf remains platform-independent, like all other audio driver headers.
- Windows crap we can't define out (`CONST`, `OUT`,  etc.) doesn't pollute anything that includes the header. Some of these names conflict with some of our names.

The cost is limited to one (1) up-front allocation. The RealTimeThread otherwise operates entirely on its own, remains alive for as long as the driver does, and is never actually dereferenced, avoiding any further overhead.

C++ gritty detail: `unique_ptr` means that the parent type gains a non-trivial destructor, but it can't generate that destructor automatically since the type is opaque in the parent class definition. This means we have to manually declare a destructor prototype, assuring C++ that we're on it; and then define it only in the implementation file, where the `unique_ptr`'s type is now known, and its destruction routine can thus be generated.
